adding the profile type management

+++ app.py
@@ -49,7 +49,7 @@ def create_user():
     password = data.get("password")
     
     if username and password:
-        user = User(username=username, password=password)
+        user = User(username=username, password=password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado com sucesso!"})
@@ -72,6 +72,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida!"}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -84,13 +87,15 @@ def update_user(id_user):
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+    if current_user.role != "admin":
+        return jsonify({"message": "Operação não permitida!"}), 403
 
     if id_user == current_user.id:
         return jsonify({"message": "Deleção não permitida"}), 403
     
     if user:
         db.session.delete(user)
-        db.session.commit()
+        db.session.commit() 
         return jsonify({"message": f"Usuário {id_user} deletado com sucesso!"})
         
     return jsonify({"message": "Usuário não encontrado"}), 404
